Removed lazy-loading feature
Although this may seem to be a controversial decision, our opinion is that the view-model should only contain the values needed for the view.
Offering the option to make properties to be available lazily, suggests they are optional, questioning whether they should be part of the view-model.

Having not used the lazy-loading feature on our own projects, there feels to be an additional overhead by checking for the `LazyLoad` attribute and `virtual` properties in Refection.